+++ src/main/scala/com/raquo/laminar/keys/ReactiveProp.scala
@@ -1,36 +1,34 @@
 package com.raquo.laminar.keys
 
 import com.raquo.airstream.core.Observable
-import com.raquo.domtypes.generic.Modifier
 import com.raquo.domtypes.generic.codecs.Codec
 import com.raquo.domtypes.generic.keys.Prop
 import com.raquo.laminar.DomApi
-import com.raquo.laminar.api.Laminar.{HtmlElement, optionToModifier}
-import com.raquo.laminar.setters.Setter
+import com.raquo.laminar.api.Laminar.{HtmlElement, optionToSetter}
+import com.raquo.laminar.modifiers.{KeySetter, Setter}
+import com.raquo.laminar.nodes.ReactiveElement
 
 class ReactiveProp[V, DomV](
   override val name: String,
   override val codec: Codec[V, DomV]
 ) extends Prop[V, DomV](name, codec) { self =>
 
-  @inline def apply(value: V): Modifier[HtmlElement] = {
+  @inline def apply(value: V): Setter[HtmlElement] = {
     this := value
   }
 
-  def maybe(value: Option[V]): Modifier[HtmlElement] = {
+  def maybe(value: Option[V]): Setter[HtmlElement] = {
     value.map(v => this := v)
   }
 
-  def :=(value: V): Modifier[HtmlElement] = {
-    new Setter[ReactiveProp[V, DomV], V, HtmlElement](this, value, DomApi.setHtmlProperty)
+  def :=(value: V): Setter[HtmlElement] = {
+    new KeySetter[ReactiveProp[V, DomV], V, HtmlElement](this, value, DomApi.setHtmlProperty)
   }
 
-  def <--($value: Observable[V]): Modifier[HtmlElement] = {
-    new Modifier[HtmlElement] {
-      override def apply(element: HtmlElement): Unit = {
-        element.subscribe($value) { value =>
-          DomApi.setHtmlProperty(element, self, value)
-        }
+  def <--($value: Observable[V]): Setter[HtmlElement] = {
+    Setter { element =>
+      ReactiveElement.bindFn(element, $value) { value =>
+        DomApi.setHtmlProperty(element, self, value)
       }
     }
   }